streamlit-frontend/requirements.txt
streamlit==1.32.0
httpx==0.27.0
pandas==2.1.4
python-dotenv==1.0.0 